complete some project functions

+++ src/main/java/com/mobileshop/entity/UsersRolesEntity.java
@@ -1,5 +1,6 @@
 package com.mobileshop.entity;
 
+import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -14,7 +15,7 @@ import javax.persistence.Table;
 public class UsersRolesEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "roleid")
 	private RoleEntity role;
@@ -22,11 +23,11 @@ public class UsersRolesEntity {
 	@JoinColumn(name = "userid")
 	private UserEntity users;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
